04 Creating a Page Layout Component
According to NextJs we have pages dir where _app.js file and in there we export MyApp function and passing in there the Component and spread pageProps.
Also, in the pages dir we have _document.js file where we export a class MyDocument that extends Document from NextJs. In there we render and return an Html element and also use some useful NextJs elements.