05 Creating our Header and Nav Components
At components dir we crate a new Header component and use it in Page component. Also, another feature that NextJs does, we want to be able to move from page to page as fast as possible, without actually having to refresh the entire page. It uses a combination of HTML5 push-state which is changing the URL as well as the ability to trigger and re-render pages when that URL has changed. So what we'll do is instead of using a href in a-html-tag, we use the Next's Link component. So that no reloading of the page happens, means faster navigation respond.
At components dir we create a Nav function-component where in there we return a nav-html-component with  next-link-components that linking to every page of the app.