06 An Intro to Styled Components and CSS
Benefits of using this library:
First, we're gonna be able to scope our CSS really easily.
Second, you can use JavaScript values inside of CSS.
The minus of the Styled Components is that it makes a flickering on page refresh. To solve this we using server-side rendering.

We define a const Logo var as a styled.element and in there we define the css we want to apply at. In order to use this styling we embed the html content inside this Logo element. The element transpile itself to a html elements with classes that have the styling.