07 Global Styles, Typography and Layout Styles
Some really basic stuff like colours, fonts and sizing of items that all needs to happen globally. So, to do that we define a GlobalStyles var in the main Page.js file and give it the Styled Components createGlobalStyle and put any styles that we want to be globally applied. In there define @font-face, In there define css-vars so that to define RGB Colours, max-width-sizes, etc. Now in order to just use global styles go into your page component and inject them right in.
Also, you could define InnerStyles var as a styled-element-object. In there put the styling of html-content.