08 Visualizing Route Changes
If the website does get slow, whether it's slow internet connection then the user needs to know that something is happening. To do that, we're going to implement this NProgress. Its the slim progress bar that goes full from left side to right, when you click on one of the URLs. At the _app.js file and we're going to implement it. We need the NProgress.js package and we're gonna start it as soon as a link starts and finish it when the page transition is done. We use the Next's Router on event route-change-start then call a callback function ()=>NProgress.start(). Also for the event of route-change-complete. Also for the event of route-change-error.