09 Fixing Styled Components Flicker on Server Render
In order to fix the about-a-second-latency use the documentation from the https://styled-components.com/docs/advanced#server-side-rendering
At the _document.js file in the class MyDocument and use the styled-components object ServerStyleSheet to initialise a new sheet object. Also, crate a page var that takes renderPage() function and callback functions in App=>props=>sheet.collectStyles() And then inside there  <App {...props}  />
So what that will do is it will go ahead and go through every single component that needs to be rendered out to the page and it will see if there's any style components in those components.