11 Setting up MongoDB, Keystone and Typescript
We need to put username/password in a .env file in a local development system, so that they will not exposed in Version Control System. When we deploy, we'll figure out how to replicate all of the secrets as environmental variables. By default, MongoDB will only allow connections from specific IPs.
Also defining some of the secret keys in the keystone.ts