13 Creating our first User data type
The User Schema have the main fields of name, email and password. Using a named export, and not an export default is because it makes auto imports of those. Needs to reboot the development with  npm run dev  for the changes take effect with generating schema. KeystoneJS Dashboard then, will have Users interface, in order to create new user.