14 Adding Auth to our Application
For Login and Role control of the Users will be used  withAuth  from  @keystone-next/auth  and feed this object properties with the User's fields.
We make a control to show UI only for people who pass the check of  isACCessAllowed  where we pass the session of the current User. Then the development needs a reboot with  npm run dev  in order the changes to take effect.