16 Uploading Product Images
Use Cloudinary direct on your website or you could just use Cloudinary for uploading images and then use the Next image tag to actually display them. We create an new data type called ProductImage and then we have a relationship of that ProductImage with the actual product.
We create a  ProductImage.ts  file to define the new Schema, so that to have the ability to upload an image of a new product. The image goes at the cloudinary folder of the app.