17 Creating two way data relationships in Keystone
The whole idea behind two-way data relationships is you should be able to start at any piece of data and get all the related data for that.   Relationships between pieces of data: if we have a product image, we should be able to figure out what products use that image. And then if we have a product, what image is associated with it.