19 Setting up Apollo Client
As we start to pull in data from our API into our front end of our application, so in order to make those queries, cache the data, evict those items from the data, handle mutations, give us the data that comes back from the server, Apollo Client. For fetching and pushing data to our API, so what they do is they just wrap it up into the package called Apollo Boost. In order for us to do image uploads, we need to use apollo-upload-client. NextJs  needs to go and fetch all of the queries that are in all of the children components. To do that we implement async function MyApp.getInitialProps.