22 A real good lesson in React Forms and Custom Hooks
At Components we create the  CeateProduct.js where implementing form input data. In NextJs new component we don't need to import React library, it's assumed. We are going to use  useState  hook, which is reactive, meaning that when that piece of state changes, it will be updated wherever it's used. So we have the ability to sync some state of inputs and have it all in a single object.
At lib folder we create  useForm.js custom hook and implement it. After that we go at CreateProduct and change the hook useState with the custom hook useForm. In react-dev-tools you can see details of hooks in Components tab.