24 Creating Products via our Mutations
In CreateProduct component and defining gql  mutation CREATE_PRODUCT_MUTATION(){}  and passing in which variables are getting passed in and what types are they. In the body of the mutation there is  createProduct(){}  where we are giving the above variables to data object properties.
Then, we create a hook from Apollo Client  useMutation()  passing CREATE_PRODUCT-MUTATION and also passing any additional data that needs to come along. We're going to destructure it in an array that it returns first the actual function that will run the mutation  createProduct  the second thing that it returns is the same thing as the query, an object {loading,error,data}. The  createProduct  is a function that is bound to this whole mutation and when we run it with the arguments it will go off to the backend.