27 Displaying Single Items, Routing and SEO
CRUD means create, read, update, and delete, CRUD-S which would be for subscribe.
Next.js used to not have file base routing, but now it does. So we're able to put things like the product ID right into the URL, or you could go even one step further and you could generate a slug for the product. To do that going to pages folder and create a product folder with a file in it [id]  Now have the ID being supplied to us because we name the file square bracket id.js and then what that does is match anything of this URL product/whatever