29 Using useEffect to deal with a tricking loading state issue
In useForm.js we are going to use useEffect where we pass a function that runs when the things we are watching change.

It updates state to initial object, which causes a change. We go into useForm.js, an do this, when the initial data changes, like we go from loading to actually having the data, then we need to update it, and the way to do that is with useEffect. It will allow us to monitor pieces of state and then run some code.