30 Deleting Products
Implement the delete function of the product, so that, when you click Delete button, it will remove the item from the product list. There's actually two ways you could go about this, one is to delete the item entirely and the other is to change the status of the item as a soft delete. In Delete Product component we pass children and export them and the reason  is that we can literally put anything we want in the component where we use it, make it more flexible. Then we need to write a mutation that will actually delete the item from the database,  the Delete Product Mutation.