31 Evicting Items from the Apollo Cache
What we're hitting now is when we delete an item it actually doesn't disappear until you refresh the page, and that's because that item's deleted from the database but it still exists in memory. So, Apollo 3 has a really nice API called Evict. React will notice that the cache change and it will re-render and take that item on out.
We create an update function that we pass along with useMutation. The update function gives us access to the cache, as well as the payload. The payload is what gets returned from the update of the mutation. So, the way that we can remove that item from the cash is by run cache.evict and then we need to pass it referenced to the actual item. The way that you can find the item in the cache is cache.identify and pass it  on payload.