32 Pagination Links
First we have to render the actual links. We create a function pagination that will take in what page we are currently on and then from that, in order to calculate. So to get the total number of items, Keystone actually provides us with meta queries. They start with underscore, it's all products meta, and then in there, really the only thing that you can query is the count. So now if we run that, we get how many products are in the database.
As about next, if you're going to be putting things like aria-disabled in Link components, then you have to nest an anchor link inside of it.