33 Pagination Dynamic Routing
We are going to make a Folder, and we're going to call it Products we need a New File that is going to be [page].js this is what the variable will be called when it's passed to you via the query param. This is file based routing. So, what's happening here is that we want the same thing to render out on three different URLs.
File-structure Restructuring by moving products.js file in the products folder.