34 Adjusting Query for Pagination Values
We go into products and we need to modify this query to not give us all the products but we want to get the first four products or two products and then we also need to offset them based on which page we currently are on.
So let's take that page and we'll pass the page to our products, and then we'll go over to our products query and we need to take in a skip variable that is going to be an integer that defaults to zero. So the Skip value is going to be calculated based on what page we are on. Then the other variable that we need to pass is called first, that is a type of integer.
So Skip is equal to page and then we multiply that by per page, that value needs to be imported from our config file, and then we also need to subtract per page because we want the lower end of it.