35 Custom Type Policies and Control over the Apollo Cache
What happens when you delete an item on pagination? How Apollo is currently handling the cache? So the idea here is that when we delete an item, what needs to happen is that item needs to be deleted from the cache and all the other items need to come sliding forward. Apollo have this in-memory cache, widthdata.js
 When you delete an item what happened is that it automatically go and fetch the next item for the next page. That's because Apollo cache, all of our items are now under one key which is all products and they're not under multiple keys which is page one, page two, page three, page four.