36 Querying the Current User - Registration & Authentication
We go to Nav.js and create the currently logged in user in the navigation. if you are not logged in, we're gonna show a link here to sign up or sign in page.
Create User.js component for the current user, as well as a custom hook that will allow us to get the current user. we can create a custom hook useUser() that will just return either the current user or not by calling CURRENT_USER_QUERY GraphQL.