37 Creating a Sign In Component
Create SignIn.js component where we need a form that is going to allow us to get your email address and password.
export default function SignIn( ) { }
Make a function called handleSubmit that takes in an event but  is this security issue because the password is in clear text?  The password is submitted over the network to your server in clear text but then  the server hashes it and compares it to what it has on file for its own hash. So, that's how all passwords work on the internet, we use HTTPS on our websites, and what HTTPS does is that when data goes between the client and the server, HTTPS make sure that that's encrypted, meaning that anyone that could possibly sniff these packets along the way, the password is encrypted.