38 Creating a Sign Out Component
Make a SignOut.js component in the URL bar, and  when you click it, it should sign you out and then re-render the navigation so, it shows you the signed out state.
Implement SIGN_OUT_MUTATION GraphQL query where calling endSession.
In order to debug it, at the button element we temporary change this code to this:
onClick={signout}   >>---=>   onClick={()=>{console.log('Signing Out ...');signout()}}
and just put a little inline function and log it to console.