40 Password Reset - Requesting a Reset
When you request a password reset, the back-end of the application generates a random token, then you are emailed a link with that token in the URL, and then when you try to reset the password, you pass it, your email address and new password and the token. Keystone has all the flow for that but it doesn't send email.
Create RequestReset.js component that calls REQUEST_RESET_MUTATION GraphQL query. At the keystone.ts of the backend