42 Password Reset - sending the email
In order to impliment the emailing of the password reset token to the user. Going into lib folder creating a new file called  mail.ts  and then we create a transporter. It is going to allow us to hook up to an SMTP API, and send out emails. So we're gonna use Ethereal fake SMTP service and we go ahead and create Ethereal Account and then it just gives you a username and a password. So we need to put this username and password in our environmental file.