43 Cart - Creating the Data Type and Two Way Relationships
As about shopping Cart Development, you gonna be able to add things to cart and display what's in it. In order to do that we work on the backend to create a new data type, the cart item. A new Keystone list  CartItem  where the fields are the quantity an integer type, product a relashionship refers to the product, user a relashionship refers to   User.cart
So what we do is a new data type called  CartItem  then, we update  User  and then we import that  CartItem  into schema.
A restart of the backend part of the application must be done for the schema changes of the database to take effect.