43 Cart - Displaying Items in a Custom Component
As of Shopping Cart Development now we make a new component called  Cart.js  and implement it, cart items that the user has. We go into this  useUser  function and update query because this now have information about our Cart. We use the Supreme CSS styles. Also, we define the styles of the list  styled.li` `  give them to  CartItemStyles.
At  CartStyles.js  styled component passing a prop to the style component called open to do this: if the props that open is true then transform translateX 0%, because the default is translate X 100% which put it off screen and then when it has a prop of open it will slide on in.
Go into lib folder, make a new file  calcTotalPrice.js  that takes in cart and return  cart.reduce( )
Tip: we'll need a key, 'cause anytime mapping in React always have to give it a unique key to that component.
Tip: because the initial renders on the server we take a look at the console of the development server.