45 Cart - Using React Context for our Cart State
We want to open and close the shopping cart slide window part using React Context, the local state in a React application. So, to be able to open the cart by pressing a button in navigation panel or maybe you also want it to open the cart when somebody adds an item. In React need to define that functionality at a high level and then you can pass it down. We put it in the Context state that allows us to define data. It is technically going to be both a component, which is a provider and it's going to be a custom hook as well.
So there's gonna be two parts to creating  cartState.js  One that's a provider that will store things like pieces of states as well as updater functions. Then, in order to access those pieces of states and updater functions anywhere else in application by calling a consumer.
We have  provider and consumer, the hook useContext is the consumer end of it. So what we did is we defined our CartState so we can use it anywhere by simply define   data=useCart( )  Then we take data and de-structure it into   {cartOpen}=useCart()  so we can pass as attribute  cartOpen  value to component
<CartStyles open={cartOpen}>   </CartStyles>
So now, we have the functionality of the shopping cart by open-close as true-false changing the cartOpen variable. In order to not expose  cartOpen  variable, also not expose any other functions. This should actually live in a piece of state. To do that, we change it to
const [cartOpen, setCartOpen] = useState(false)
Also, we can create helper function  toggleCart()  we'll  setCartOpen  to be whatever the opposite.