46 Cart - Adding Items to Cart
Write a mutation that is called  createCartItem. So, basically take the existingGraphQLSchema,
Make a new file called  index.ts  in mutations folder, where we export   extendGraphQLSchema  a  graphQLSchemaExtension({})  with  typeDefs:graphql``  and  resolvers:{}   in it.
Make a new file called  addToCart.ts  in mutations folder, where we export function  addToCart  We make it an
async function addToCart(    ): Promise<CartItemCreateInput>  that returns a promise of this type.
Then we implement: Query current user to see if they are signed in, Query the current user's cart, See if the current item is in their cart, If it already been then increment by one for more items of this, If it isn't create a new shopping cart.