48 Cart - Animating the Cart Cart Value
Make a new component cart count that exports default function  CartCount()  that takes in the number of items in the cart and returns  <Dot>{count}</Dot>  a styled.div``  CSSheet in Javascript. We nest  <Dot> inside React  <TransitionGroup><CSSTransition>  components . We need to implement this into  Nav.js  Basically, at  CartCount  component we pass prop
count={user.cart.reduce((tally,cartItem)=>tally+cartItem.quantity,0)}    So now, if we add an item to our cart it displays the quantity of items in the cart. So, when you add an item to cart, by adding and removing those classes, we'll be able to hook onto them with CSS and do some kind of neat CSS animation.