51 Search
Go and make a new component  Search.js  which export default function Search() and then from that, we're going to return <SearchStyles> and inside it <DropDown> and inside it multiple <DropDownItem>  So that, each search result that comes back will display in lines of a dropdown menu.
Then we're going to use a library called Downshift, which makes this sort of dropdown, auto-complete stuff easily and to be accessible, making these sort of drop-downs accessible for keyboards.
In order to debouncing the function that goes searching, means to fire off but hold on for a couple seconds. So, we'll take this findItems and we'll make it into a debounce function from lodash.