54 Creating Order and OrderItem Data Types
We need to create an order type, as well as order items, in which copy all the fields from the product into an order item. The id is not actually a really good thing to show, so what we can do, is to put a label on this field, so the label field is going to be a virtual field, a virtual field is something that is calculated on demand that's stored in the database, but it's something that is just generated as we need it. Whatever returns this resolver will be what is shown up as the label. In Keystone dashboard we can modify what columns of the Orders schema will be output.