59 Displaying a Single Order
As about frontend Order Displaying, creating the single order page to display the content of this page and and the total of the items, total money of order, etc. Also, there should be a page that lists all of these orders. So we need to do NextJS Routing of   page/order/id  To do this going at pages and create a new order folder with file  [id].js  in it that catch this specific route. We export default function  SingleOrderPage()  We write  SINGLE_ORDER_QUERY  that gets from  gql`query`