60 Displaying All Orders
As about frontend Order Displaying, we already have all  orders.js  page, so going to re-constract it like the single order page, as the previous commit.
All Orders is a page of a list of already done orders, the quantity of stuff and the total price.
Tip: sometimes if you're using the ID for something, and then you're using that ID somewhere else on the page as well React is going to yell at you!
Err: Unhandled Runtime Error: ... expected to receive a single React element child.
because we're stuffing everything in a <Link></Link> this throws error, So we can wrap everything in it, in a React fragment <></> element.
Tip: another improvement we should do is an anchor link in the previous tip's fragment <a> element.