63 Basic Access Control via Sessions
As about Roles, Permissions and Restricting Access, we're going to have access to all of these  permissions on the session. The session is a little bit of information about the currently logged-in user that we get access to when we are doing server side check. Keystone has access to the session when it checks if you're able to update an item or delete an item or to create an item or to update a specific field. So what we need to do is to   keystone.ts  configuration, that we query on every session for the user's id, the name and the email of the currently logged in user. And what that does is every single time there's a request, Keystone goes and grabs that data from the database, like  addToCart.ts  mutation. So what we want to do is to query the user's roles including the id, the name, and the email on every session because we're going to need these pieces of information when we go to give or not give permition to whether you try to access something.
These are the CRUD Database Operations.
Create a new file  access.ts  in the root of the Backend folder to be together with  types.js   keystone.ts  where we implement access control functions. So that, at the resources of the database go and update the  keystone/schema list({access})  with new access object with properties: create, read, update, delete giving initial value false.