65 More Flexible Rule Based Functions
As about Roles, Permissions and Restricting Access, at the access typescript file, at the end of that we impliment rule-based functions.
Rules can return a Boolean Yes or No or a filter which limits which products they can access. So export rules is an object
 1.  Do they have the permission of canManageProducts?
 2.  If not, do they own this item?
  These are  logical functions that can be used for list access.
Example: who should be able to update and delete a product, the creator of the product should be able to update and delete their own products without having an elevated privilege of managing products, either you have canManagedProduct role or you are the owner of the item.