67 Cart and Order based Rules
As about Roles, Permissions and Restricting Access, these are three different lists:  the cart,  the orders and the cart items but they all have the exact same rules.
At  backend/schemas/OrderItem.ts  in the  list({access})  we have a new  read:rules.canManageOrderItems   At  backend/access.ts  we duplicate the rule:
  canOrder({session}:ListAccessArgs){<--->}
  canManageOrderItems({session}:ListAccessArgs){<--->}
but we're not going to filter it where the order items user is, we're going to filter it where the
  return{order:{user:{id:session.itemId}}}
So, we take the orderItem, we query the order and then we query the related user and then we check their id to make sure that it lines up with what they have.
When you are in a signed out state, if you try to query all the orders then you don't get access to them. But it does throw an error, it should give us a not like an internal server error, it should give us like an access is not defined. So to solve this is to the  backend/access.ts  at this rule adding an if-statement
  canOrder({session}: ListAccessArgs){if( !isSignedIn({session}) )return false}