70 Creating a Gated Sign In Component
As about Roles, Permissions and Restricting Access,  go at  frontend/components/PleaseSignIn.js  and make a gated sign-in component, which will render out the sign-in component, where  export default function({children}) {useUser()}  takes in children prop and use the useUser function from User component and check if there is not exist me variable, means you are not signed in, we return SignIn component. If you are signed in, we'll  return the children.