71 Test Setup, Tooling and Methodology
As about Testing, we are using Jest, a test runner, It's just about run it, see what comes back,
and React Testing Library, a DOM testing library, it's going to take all of our React components,  given the data, it renders and gives the output expected to the user.
At the  frontend/jest.setup.js  we  import '@testing-library/jest-dom'  and what this does is it allows the test runner, Jest know about the testing library additions, and because we are going to be running these tests in a headless mode. Also configure this  window.alert=console.log