72 Testing Basics
As about Testing, is how to describe and write a test case expected something to equal something else. Tests work by writing a file that has a bunch of tests that are related to one thing.
Every  Component.js  file will have its own  Component.test.js  file. In this app we are building, Jest tests live in  __tests__  where creating  file.test.js  where test cases in this test block
  describe('This is the test we perform to', ()=>{it('work as expected', ()=>{expect(3).toEqual(3)})})
In each test we can write one or many expect statements to see if the expected test case will pass. So that make sure that the function or the component, is outputting and returning the right values  we expect.
In terminal  npm run test  which is an alias of  jest --watch  that go find tests in place.
expect().toBe, toBeCalled, toBeCalledWith, toBeChecked,