74 React Testing Library
As about Testing, going at product component that takes in a product and displays the product photo, the name, the formating price, and a price tag wraps a link around it, and then has a bunch of buttons to add and delete the items. So, going at tests make a  product.test.js  file and describe what we like to do  <Product/>  component with tests. In order to do that, we need render function from React Testing Library.
Going to use utils from  frontend/lib/testUtils.js  file ...
So if I try to render this out with a fake product, and then we pass the product attribute in Product component
  const product = fakeItem()
also in describe block
  const {container,debug}= render(<Product product={product} />)
If we take a look at our  frontend/pages/_app.js  file we have wrapped the entire application in Apollo provider and then we render out the actual component that we have. So we need to do something similar to that in our tests, but we don't want to render out the entire application because then that would mean having to connect to a database and those tests are very flaky and they're not actually testing the actual component. So because of that, Apollo provides us with a mocked provider that is going to allow us to fake having Apollo in there as well as fake some of the API responses for the data that it needs. So, we import MockedProvider from Apollo React Testing and then we take this MockProvider we  wrap it around every single component in component testing.