75 Snapshot Testing
As about Testing, with react testing library is that we use screen function to render, then we check for specific things. So the idea with snapshot testing, is basically you take a picture or a snapshot of what the HTML looks like for each component and then when those change they will fail and you'll know that something has changed. We've got a folder  __snapshots__  open that and go into  product.test.js.snap  file. Now, what happens, if I were to go to  Product component and make changes at the layout, when tests runs it will find mismatches with the snapshot. So Jest runner says that we have to inspect code changes, or press u for update, it goes and updates that snapshot. We can just take a snapshot of one piece of HTML and expect that to look exactly the same as we go on.