77 Mocking GraphQL Data Requests
As about mocking data in testing, we need to be able to query inside of our tests. In order to eliminate any databases or network requests from your tests and you want to keep them isolated. Meaning that you stub out the network request for some data that you have just locally generated.
Go at   SingleProduct.js  file we have a  SINGLE_ITEM_QUERY  inside of that we query the item based on its ID. Then inside of that, we do things like render out the ProductStyles, with the Head, the image, and the details inside of it. So, we go into tests folder create  SingleProduct.test.js  where we define mocks variable that is an array of an object that when request this query object then result this data object. Also, we describe  <SingleProduct/>  where mocking data by making Apollo query a single item based on the ID that is passed. We use render function from Testing-library where  to render <MockedProvider mocks={mocks}> nest in <SingleProduct id="123" /> pass in this id prop. It's rendering out the console.log, so what happens if you render it immediately, you're going to see the loading state, because it needs to have another pass of the render in order to do the query. So, we'll go back to  SingleProduct.js  and to  <ProductStyles data-testid="singleProduct"> and pass this prop. Then at  SingleProduct.test.js  calling screen.findByTestId() awaiting for the data-testid. The  screen.findByTestId()  is async-await based function so we have to write this test case appropriate.