78 Updating Props and re-rendering
As about testing, let's look at how to test components when the props update to make sure that they really render appropriately. We have CartCount component, which is a little circle with the number of items in your cart with a bit of an animation, to show that it goes from one to another, so there's a little bit of time. That time needs to pass before we test if things are finished.
Go at the new  CartCount.test.js  file, a new test block that describes <CartCount/>  that it renders the component passing the count prop with number 10. If there is any error that's thrown inside of  <CartCount/>  the test will automatically fail. Next, let's do snapshot of it, so it matches snapshot by giving to the container  that it renders the component passing the count prop with number 11 and expect that the container to match snapshot. Next, write test that it updates via props by giving to the container  that it renders the component passing the count prop with number 11 and expect container's text content to be 11. This last test is how you would test if something changes when it updates via prop. In order to do this we call rerender function passing CartCount component with count prop number 12, and expect the container to have 12. Because of CSSTransitionGroup from React it will temporarily have both of those values in the outgoing element, the pre-transition and after-trnsition. So what we need to do is to wait a little bit of time for it to show up.
React Testing-library have async-await utilities:  screen.findBy()  and  screen.waitFor()
Another way is to use  await wait(400)  a waiting function passing milliseconds