79 Testing Signed in and Signed out Nav States
Testing about, let's test  <Nav/>  which have three different states. It is different when you're signed in versus being signed out. This component is also different when you actually have items inside of your cart. So, all of these states of the component is going to involve mocking and then running tests against them to make sure that everything is working appropriately. New file  Nav.test.js  into test folder and define a structure of array of objects that have the request object query and the result object data. Give this structure at vars of noSignedInMocks, signedInMocks and signedInMocksWithCartItems.
In the describe block of the  <Nav/>  we have an it block that renders with minimal items when signed out.
In the describe block of the  <Nav/>  we have an it block that renders with full items when signed in.
In the describe block of the  <Nav/>  we have an it block that renders the amount of items in the cart.
Every test case of them have  <MockedProvider>  pass in its own mocks prop. We have three different types of mocks of every state of the component and there's lots of little edge cases for us to cover.