80 Pagination Testing
Testing about, we're going to move on to the pagination testing. Test utils function  makePaginationMocksFor(length)   take in a length. Make file  Pagination.test.js  where  describe('<Pagination/>')  with test that  it('displays a loading message')  by using Testing-library's  render(<MockedProvider><Pagination/><MockedProvider/>)  passing in prop  mocks={makePaginationMocksFor(1)}  for one page. From this test we  expect(container).toHaveTextContent('Loading...')  We always have to round up for our pagination.
Next, we need to make sure that the previous and next buttons are working appropriately, for your users who are going to click them as well as for accessibility.