81 Testing User Events and Mutations
Testing is about, render the component out, give it some mock data, check that it renders out properly. But what we now need to start testing is when people type into boxes and click on buttons and things like that, need to make sure that we cover those tests have to do with user interaction, that's another aspect of react testing library.
New file  SignUp.test.js  and first make a var  me  as   fakeUser() The mocks that we need have two parts, when somebody signs up for an account, we're going to have to mock comes back from that.