82 Testing Password Reset Component
Testing about, the  RequestReset.js  component, go at tests and create a new file. Well, somebody submits the form with an email address, we are mocking the request giving an email address and as a result sending a user password reset link. Then, describe the RequestReset component by saying that
 - it renders and matches snapshot
 - it calls the mutation when submitted