83 Mocking 3rd Party Libraries
This one's a little bit different because we're rendering or checking for things, we're mocking data, we're typing into boxes, but goes a little step further, we're mocking be mocking out the entire router. So when you create a product, and it's successfully created, we changed the page. So, how do we test that the page has been changed? Without actually having a page, this is sort of helpful when you're using global modules like router. So let's get into it, a test file  CreateProduct.test.js  and  describe('<CreateProduct/>')  that:
 - it renders and matches snapshot where expected the container to match snapshot
 - it handles the updating: render the form, type into the boxes, check that boxes are populated
Let's let's build another test, which is when you actually type these values, you can go ahead and submit them.
 - it creates the items when the form is submitted: create the mocks, type into the boxes, submit it and see if the page change has been called